made changes to flight-tracker and main.js
