+++ src/flight-tracker.js
@@ -1,10 +1,11 @@
 export class FlightTracker {
-  async getFlight(flightNumber, date) {
+  async getFlight(flightNumber) {
+    console.log(flightNumber)
     try {
-      fetch(`https://aerodatabox.p.rapidapi.com/flights/${flightNumber}/${date}?withLocation=false&withAircraftImage=false`, {
+      let response = await fetch(`https://aerodatabox.p.rapidapi.com/flights/DL47/2019-08-29?withLocation=false&withAircraftImage=false`, {
         "method": "GET",
         "headers": {
-          "x-rapidapi-host": aerodatabox.p.rapidapi.com,
+          "x-rapidapi-host": "aerodatabox.p.rapidapi.com",
           "x-rapidapi-key": `${process.env.API_KEY}`,
           "mode": "no-cors"
         }
@@ -15,6 +16,7 @@ export class FlightTracker {
       } else {
         jsonifiedResponse = false;
       }
+      console.log(jsonifiedResponse);
       return jsonifiedResponse;
     } catch (error) {
       return false;
